requirements.txt
fastapi
uvicorn
sqlalchemy
transformers
torch
pillow
pdf2image
python-docx
httpx
python-dotenv